+++ moor/lib/src/runtime/query_builder/statements/select/custom_select.dart
@@ -22,14 +22,6 @@ class CustomSelectStatement with Selectable<QueryRow> {
 
   /// Constructs a fetcher for this query. The fetcher is responsible for
   /// updating a stream at the right moment.
-  @Deprecated(
-      'There is no need to use this method. Please use watch() directly')
-  QueryStreamFetcher<List<QueryRow>> constructFetcher() {
-    return _constructFetcher();
-  }
-
-  /// Constructs a fetcher for this query. The fetcher is responsible for
-  /// updating a stream at the right moment.
   QueryStreamFetcher<List<QueryRow>> _constructFetcher() {
     final args = _mapArgs();
 
@@ -48,12 +40,6 @@ class CustomSelectStatement with Selectable<QueryRow> {
   @override
   Stream<List<QueryRow>> watch() {
     return _db.createStream(_constructFetcher());
-  }
-
-  /// Executes this query and returns the result.
-  @Deprecated('Use get() instead')
-  Future<List<QueryRow>> execute() async {
-    return get();
   }
 
   List<dynamic> _mapArgs() {